+++ src/renderer/i18n/de/savers.ts
@@ -1,0 +1,7 @@
+import { SaversMessages } from '../types'
+
+const savers: SaversMessages = {
+  'savers.noSavings': 'Du hast keine Savings'
+}
+
+export default savers