+++ src/renderer/i18n/ru/savers.ts
@@ -1,0 +1,7 @@
+import { SaversMessages } from '../types'
+
+const savers: SaversMessages = {
+  'savers.noSavings': "You don't have any savings - RU"
+}
+
+export default savers